refactor div selected
