+++ apps/web/client/src/app/project/[id]/_components/editor-bar/text-inputs/font-weight.tsx
