+++ apps/web/client/src/app/project/[id]/_components/editor-bar/text-inputs/text-color.tsx
